feat: lwk datasource
